+++ examples/ssr/src/pages/scroll/_components/Photos.ts
@@ -3,7 +3,7 @@ import { fadeInOut } from 'ficsjs/animation'
 import { goto, queries } from 'ficsjs/router'
 import { absoluteCenter, oklch } from 'ficsjs/style'
 import Icon from '@/components/Icon'
-import Skeleton from '@/components/scroll/Skeleton'
+import Skeleton from '@/pages/scroll/_components/Skeleton'
 import { api, getPhotos, UNIT_LENGTH } from '@/data/photos'
 import type { Photo } from '@/types'
 import { CircleX } from 'lucide-static'
