+++ examples/ssr/src/pages/scroll/_components/Skeleton.ts
